+++ src/main/java/com/onlyjavatech/samir/service/EmployeeService.java
@@ -1,5 +1,7 @@
 package com.onlyjavatech.samir.service;
 
+import com.onlyjavatech.samir.exception.ObjectNotFoundException;
+import com.onlyjavatech.samir.exception.ValidationHandler;
 import com.onlyjavatech.samir.model.DepartmentModel.Department;
 import com.onlyjavatech.samir.model.DepartmentModel.DepartmentResponseModel;
 import com.onlyjavatech.samir.model.Employee;
@@ -10,13 +12,17 @@ import com.onlyjavatech.samir.repository.EmployeeRepository;
 import com.onlyjavatech.samir.service.DepartmentService.DepartmentService;
 import com.onlyjavatech.samir.service.ProjectService.ProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 @Service
 public class EmployeeService {
@@ -31,6 +37,31 @@ public class EmployeeService {
 
     @Transactional
     public EmployeeResponseModel registerEmployee(EmployeeRequestModel request) {
+        if (request == null) {
+            throw new ObjectNotFoundException("Object not found", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getFirstname() == null || request.getFirstname().isEmpty()) {
+            throw new ValidationHandler("Please Provide Valid First Name", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getLastname() == null || request.getLastname().isEmpty()) {
+            throw new ValidationHandler("Please Provide Valid Last Name", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getEmailId() == null || request.getEmailId().isEmpty()) {
+            throw new ValidationHandler("Please Provide Valid EmailId", HttpStatus.BAD_REQUEST);
+        }
+        String regex = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,6}$";
+        Pattern pattern = Pattern.compile(regex);
+        if (!pattern.matcher(request.getEmailId()).matches()) {
+            throw new ValidationHandler("Please Provide Valid EmailId like abc@gmail.com", HttpStatus.BAD_REQUEST);
+        }
+
+        if (request.getDepartmentId() == null || request.getDepartmentId().isEmpty()) {
+            throw new ValidationHandler("Please Provide Valid DepartmentId", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getProjects().isEmpty()) {
+            throw new ValidationHandler("Please Provide ProjectList", HttpStatus.BAD_REQUEST);
+        }
+
         Employee employee = new Employee();
         employee.setFirstname(request.getFirstname());
         employee.setLastname(request.getLastname());
@@ -38,18 +69,21 @@ public class EmployeeService {
 
         UUID uuid = UUID.randomUUID();
         String uuidAsString = uuid.toString();
-
-        System.out.println("Your UUID is: " + uuidAsString);
         employee.setId(uuidAsString);
 
         Department department = departmentService.getDepartmentByDepartmentId(request.getDepartmentId());
         employee.setDepartment(department);
         request.getProjects().forEach(project -> {
+            if (project == null) {
+                throw new ValidationHandler("Please Provide Valid ProjectList objects", HttpStatus.BAD_REQUEST);
+            }
+            if (project.getProjectName() == null || project.getProjectName().isEmpty()) {
+                throw new ValidationHandler("Please Provide Valid ProjectList Name", HttpStatus.BAD_REQUEST);
+            }
             Project newProject = new Project();
             String projectId = UUID.randomUUID().toString();
             newProject.setId(projectId);
             newProject.setProjectName(project.getProjectName());
-//            projectService.registerProject(project);
             employee.addProject(newProject);
 
         });
@@ -57,63 +91,88 @@ public class EmployeeService {
         return setEmployeeResponseModel(newEmployee);
     }
 
+    public void removeProjectFromProjectsList(Project project) {
+        Employee employee = new Employee();
+//        employee.getProjects().remove(project);
+        employee.removeProject(project);
+        employeeRepository.save(employee);
+
+    }
+
     public EmployeeResponseModel updateEmployee(EmployeeRequestModel request) {
+        if (request == null) {
+            throw new ObjectNotFoundException("Object can't be null...", HttpStatus.NOT_FOUND);
+        }
+        if (request.getId() == null || request.getId().isEmpty()) {
+            throw new ValidationHandler("Employee Id can't be null or empty...", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getFirstname() == null || request.getFirstname().isEmpty()) {
+            throw new ValidationHandler("Employee First name can't be null or empty...", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getLastname() == null || request.getLastname().isEmpty()) {
+            throw new ValidationHandler("Employee Last name can't be null or empty...", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getEmailId() == null || request.getEmailId().isEmpty()) {
+            throw new ValidationHandler("Employee Email Id can't be null or empty...", HttpStatus.BAD_REQUEST);
+        }
+        if (request.getProjects() == null || request.getProjects().isEmpty()) {
+            throw new ValidationHandler("Projects list can't be null or empty", HttpStatus.BAD_REQUEST);
+        }
+        String regex = "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,6}$";
+        Pattern pattern = Pattern.compile(regex);
+        Matcher matcher = pattern.matcher(request.getEmailId());
+        if (!matcher.matches()) {
+            throw new ValidationHandler("Please provide valid email like abc@gmail.com", HttpStatus.BAD_REQUEST);
+        }
+
         String id = request.getId();
-        Employee employee = employeeRepository.findById(id).get();
+        Optional<Employee> optionalEmployee = employeeRepository.findById(id);
+        if (!optionalEmployee.isPresent()) {
+            throw new ObjectNotFoundException("employee not found...", HttpStatus.NOT_FOUND);
+        }
+        Employee employee = optionalEmployee.get();
         employee.setFirstname(request.getFirstname());
         employee.setLastname(request.getLastname());
         employee.setEmailId(request.getEmailId());
 
         request.getProjects().forEach(project -> {
-            System.out.println("-------------        +       ----------------------");
-            System.out.println("==== " + project.getProjectName() + "=====");
+            if (project == null) {
+                throw new ObjectNotFoundException("Project can't be null or empty", HttpStatus.BAD_REQUEST);
+            }
+            if (project.getId() == null || project.getId().isEmpty()) {
+                throw new ValidationHandler("Project Id can't be Empty or null", HttpStatus.BAD_REQUEST);
+            }
+            if (project.getProjectName() == null || project.getProjectName().isEmpty()) {
+                throw new ValidationHandler("Project Name can't be null or empty", HttpStatus.BAD_REQUEST);
+            }
             if (!projectService.checkProjectByProjectId(project.getId())) {
                 Project newProject = new Project();
                 String projectId = UUID.randomUUID().toString();
-                System.out.println("--- id ---" + projectId);
                 newProject.setId(projectId);
                 newProject.setProjectName(project.getProjectName());
                 employee.addProject(newProject);
             } else {
-                System.out.println("-- test---");
                 Project newProjectElse = projectService.getProjectByProjectId(project.getId());
                 employee.addProject(newProjectElse);
             }
         });
 
         request.getRemoveProjects().forEach(project -> {
-            System.out.println("-------------        +       ----------------------");
-            System.out.println("==== " + project.getProjectName() + "=====");
             if (!projectService.checkProjectByProjectId(project.getId())) {
-                System.out.println("--- remove ----");
                 throw new EntityNotFoundException();
             } else {
-                System.out.println("-- test---");
                 Project newProjectElse = projectService.getProjectByProjectId(project.getId());
                 employee.removeProject(newProjectElse);
             }
         });
 
-
         Employee updatedEmployee = employeeRepository.save(employee);
-        System.out.println("----- first name----"+updatedEmployee.getFirstname());
         return setEmployeeResponseModel(updatedEmployee);
-
     }
 
-    //    public List<Employee> getEmployee() {
-//        return (List<Employee>) employeeRepository.findAll();
-//    }
     public List<EmployeeResponseModel> getEmployee() {
         Iterable<Employee> employees = employeeRepository.findAll();
         List<EmployeeResponseModel> employeeList = new ArrayList<>();
-
-//        for (Employee employee : employees) {
-////            EmployeeResponseModel employee =new EmployeeResponseModel();
-//            EmployeeResponseModel employee_row = setEmployeeResponseModel(employee);
-//
-//            employeeList.add(employee_row);
-//        }
         employees.forEach(employee -> {
             EmployeeResponseModel employee_row = setEmployeeResponseModel(employee);
             DepartmentResponseModel departmentResponse = new DepartmentResponseModel();
@@ -126,27 +185,40 @@ public class EmployeeService {
     }
 
     public EmployeeResponseModel getEmployeeById(String id) {
-        Employee employee = employeeRepository.findById(id).get();
-
-        return setEmployeeResponseModel(employee);
+        if (id == null || id.isEmpty() || id.trim().isEmpty()) {
+            throw new ValidationHandler("Employee is can't be null or empty", HttpStatus.BAD_REQUEST);
+        }
+        Optional<Employee> optionalEmployee = employeeRepository.findById(id);
+        if (!optionalEmployee.isPresent()) {
+            throw new EntityNotFoundException();
+        }
+        Employee employee = optionalEmployee.get();
+        System.out.println(employee.getDepartment().getDepartment_name() + "---");
+        EmployeeResponseModel employeeResponseModel = setEmployeeResponseModel(employee);
+        employeeResponseModel.setDepartment(departmentService.getDepartmentById(employee.getDepartment().getId()));
+        return employeeResponseModel;
     }
 
     public void deleteEmployee(String id) {
-//        Employee emp = employeeRepository.findById(id).get();
-        Project project = new Project();
-        project.setId(id);
-        Employee employee = new Employee();
-        employee.removeProject(project);
+        if (id == null || id.isEmpty() || id.trim().isEmpty()) {
+            throw new ValidationHandler("Employee is can't be null or empty", HttpStatus.BAD_REQUEST);
+        }
+        Optional<Employee> optionalEmployee = employeeRepository.findById(id);
+        if (!optionalEmployee.isPresent()) {
+            throw new ValidationHandler("Employee not found ,please provide valid Employee Id..", HttpStatus.BAD_REQUEST);
+        }
         employeeRepository.deleteById(id);
     }
 
-//    public EmployeeResponseModel deleteEmployee(EmployeeRequestModel request){
-//        Integer id =request.getId();
-//
-//        EmployeeResponseModel employeeResponseModel = new EmployeeResponseModel();
-//        employeeRepository.deleteById(id);
-//        return  employeeResponseModel;
-//    }
+    public void removeProjectsRelation(Project project) {
+//        Employee employee = new Employee();
+        List<Employee> employeeList = employeeRepository.findByProjects_Id(project.getId());
+        employeeList.forEach(employee -> {
+            employee.removeProject(project);
+            employeeRepository.save(employee);
+        });
+
+    }
 
     private EmployeeResponseModel setEmployeeResponseModel(Employee employee) {
         EmployeeResponseModel response = new EmployeeResponseModel();
